requirements.txt
flask==3.0.2
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.0.2
bcrypt==4.1.2
python-dateutil==2.8.2 